For #660: rename all root test suite to AllTests
