+++ sharding-jdbc-core-spring/sharding-jdbc-core-spring-namespace/src/test/java/io/shardingjdbc/spring/AllTests.java
@@ -44,5 +44,5 @@ import org.junit.runners.Suite.SuiteClasses;
         MasterSlaveNamespaceTest.class,
         ShardingNamespaceTest.class
     })
-public class AllSpringTests {
+public class AllTests {
 }
